test: review fixes
