+++ qa/c8-orchestration-cluster-e2e-test-suite/tests/common-flows/identity-users-flows.spec.ts
@@ -11,7 +11,8 @@ import {navigateToApp} from '@pages/UtilitiesPage';
 import {expect} from '@playwright/test';
 import {captureScreenshot, captureFailureVideo} from '@setup';
 import {relativizePath, Paths} from 'utils/relativizePath';
-import {generateRandomStringAsync} from '../../utils/randomString';
+import {createTestData} from '../../utils/constants';
+import {sleep} from '../../utils/sleep';
 
 test.describe('Users Page Tests', () => {
   test.beforeEach(async ({loginPage, page}) => {
@@ -30,27 +31,27 @@ test.describe('Users Page Tests', () => {
     identityUsersPage,
     identityHeader,
   }) => {
-    const randomString = await generateRandomStringAsync(3);
-
-    const TEST_USER = {
-      username: 'Test' + randomString,
-      name: 'Test User' + randomString,
-      email: `test${randomString}@test.com`,
-      password: `test${randomString}`,
-    };
-    await identityUsersPage.createUser(TEST_USER);
+    const testData = createTestData({user: true});
+    const testUser = testData.user!;
+    await identityUsersPage.createUser({
+      username: testUser.username,
+      password: testUser.password,
+      email: testUser.email!,
+      name: testUser.name,
+    });
 
     await expect(identityUsersPage.usersList).toBeVisible();
-    await expect(identityUsersPage.usersList).toContainText(TEST_USER.username);
-    await identityUsersPage.deleteUser(TEST_USER);
+    await expect(identityUsersPage.usersList).toContainText(testUser.username);
+    await identityUsersPage.deleteUser(testUser);
     await identityHeader.logout();
+    await sleep(2000);
     await expect(page).toHaveURL(
       `${relativizePath(Paths.login('identity'))}?next=/identity/`,
     );
 
     await test.step(`Deleted user cannot access Identity`, async () => {
       await navigateToApp(page, `identity`);
-      await loginPage.login(TEST_USER.username, TEST_USER.password);
+      await loginPage.login(testUser.username, testUser.password);
       await expect(page).toHaveURL(new RegExp(`identity`));
       await expect(loginPage.errorMessage).toContainText(
         /Username and [Pp]assword do(?: not|n't) match/,
@@ -59,7 +60,7 @@ test.describe('Users Page Tests', () => {
 
     await test.step(`Deleted user cannot access Tasklist`, async () => {
       await navigateToApp(page, `tasklist`);
-      await loginPage.login(TEST_USER.username, TEST_USER.password);
+      await loginPage.login(testUser.username, testUser.password);
       await expect(page).toHaveURL(new RegExp(`tasklist`));
       await expect(loginPage.errorMessage).toContainText(
         /Username and [Pp]assword do(?: not|n't) match/,
@@ -68,7 +69,7 @@ test.describe('Users Page Tests', () => {
 
     await test.step(`Deleted user cannot access Operate`, async () => {
       await navigateToApp(page, `operate`);
-      await loginPage.login(TEST_USER.username, TEST_USER.password);
+      await loginPage.login(testUser.username, testUser.password);
       await expect(page).toHaveURL(new RegExp(`operate`));
       await expect(loginPage.errorMessage).toContainText(
         /Username and [Pp]assword do(?: not|n't) match/,
